train on sample data

+++ models/DenseNetork/models.py
@@ -26,7 +26,7 @@ class VecDataSet(Dataset):
 
     @classmethod
     def from_df(cls, path_to_dataframe):
-        x = torch.from_numpy(pd.read_csv(path_to_dataframe).to_numpy())
+        x = torch.from_numpy(pd.read_csv(path_to_dataframe).to_numpy()).to(torch.float32)
         return cls(x)
 
     @classmethod
